# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Student {
  id: number;
  nationalCode: string;
  fullName: string;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  schoolName?: string | null;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}

export interface LoginInput {
  /**
     * @minLength 10
     * @maxLength 10
     */
  nationalCode: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  schoolName?: string | null;
}

export interface AuthResult {
  student: Student;
  isNew: boolean;
}

export interface DailyRecord {
  id: number;
  studentId: number;
  recordDate: string;
  /** @nullable */
  mathScore?: number | null;
  /** @nullable */
  physicsScore?: number | null;
  /** @nullable */
  chemistryScore?: number | null;
  /** @nullable */
  biologyScore?: number | null;
  /** @nullable */
  anxietyLevel?: number | null;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  focusDrop?: boolean | null;
  /** @nullable */
  socialInteraction?: number | null;
  /** @nullable */
  examDaysLeft?: number | null;
  /** @nullable */
  freeText?: string | null;
  /** @nullable */
  riskLevel?: string | null;
  createdAt: string;
}

export interface DailyRecordInput {
  recordDate: string;
  /** @nullable */
  mathScore?: number | null;
  /** @nullable */
  physicsScore?: number | null;
  /** @nullable */
  chemistryScore?: number | null;
  /** @nullable */
  biologyScore?: number | null;
  /** @nullable */
  anxietyLevel?: number | null;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  focusDrop?: boolean | null;
  /** @nullable */
  socialInteraction?: number | null;
  /** @nullable */
  examDaysLeft?: number | null;
  /** @nullable */
  freeText?: string | null;
}

export interface ExtractTextInput {
  text: string;
}

export interface ExtractedValues {
  /** @nullable */
  anxietyLevel?: number | null;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  mathScore?: number | null;
  /** @nullable */
  physicsScore?: number | null;
  /** @nullable */
  chemistryScore?: number | null;
  /** @nullable */
  biologyScore?: number | null;
  /** @nullable */
  examDaysLeft?: number | null;
  summary: string;
}

export interface AnalysisResult {
  id: number;
  recordId: number;
  /** @nullable */
  ruleEngineOutput?: string | null;
  /** @nullable */
  interventionOutput?: string | null;
  /** @nullable */
  reportOutput?: string | null;
  /** @nullable */
  nlpOutput?: string | null;
  /** @nullable */
  riskLevel?: string | null;
  createdAt: string;
}

export interface RunAnalysisInput {
  recordId: number;
}

export interface Question {
  id: number;
  subject: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  difficulty?: number | null;
  text: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  /** @nullable */
  explanation?: string | null;
}

export interface QuestionInput {
  subject: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  difficulty?: number | null;
  text: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  /** @nullable */
  explanation?: string | null;
}

export interface GenerateExamInput {
  count?: number;
}

export interface ExamAnswer {
  questionId: number;
  chosen: string;
}

export interface ExamSubmitInput {
  sessionId: string;
  answers: ExamAnswer[];
}

export interface SubjectScore {
  subject: string;
  total: number;
  correct: number;
}

export interface ExamResult {
  sessionId: string;
  totalQuestions: number;
  correctCount: number;
  score: number;
  subjectBreakdown: SubjectScore[];
}

export interface ExamSession {
  sessionId: string;
  score: number;
  totalQuestions: number;
  correctCount: number;
  createdAt: string;
}

export interface StatsSummary {
  /** @nullable */
  avgMath?: number | null;
  /** @nullable */
  avgPhysics?: number | null;
  /** @nullable */
  avgChemistry?: number | null;
  /** @nullable */
  avgBiology?: number | null;
  /** @nullable */
  avgAnxiety?: number | null;
  /** @nullable */
  avgSleep?: number | null;
  totalRecords: number;
  /** @nullable */
  lastRiskLevel?: string | null;
  /** @nullable */
  weakestSubject?: string | null;
}

export interface RiskTrendPoint {
  date: string;
  riskLevel: string;
  /** @nullable */
  anxietyLevel?: number | null;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthResult {
  success: boolean;
  token: string;
}

export interface AdminSettings {
  /** @nullable */
  adminPassword?: string | null;
  aiModel: string;
  minResponseLength: number;
}

export interface AdminSettingsUpdate {
  /** @nullable */
  adminPassword?: string | null;
  /** @nullable */
  aiModel?: string | null;
  /** @nullable */
  minResponseLength?: number | null;
}

export type AdminOverviewRiskBreakdown = {
  green: number;
  yellow: number;
  orange: number;
  red: number;
};

export interface AdminOverview {
  totalStudents: number;
  totalRecords: number;
  riskBreakdown: AdminOverviewRiskBreakdown;
  recentStudents: Student[];
}

export type ListRecordsParams = {
limit?: number;
offset?: number;
};