# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login or register with national code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in student
      responses:
        "200":
          description: Current student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /analysis/{recordId}:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get full AI analysis for a record (cached)
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/run:
    post:
      operationId: runAnalysis
      tags: [analysis]
      summary: Run AI analysis for a record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/questions:
    get:
      operationId: listQuestions
      tags: [admin]
      summary: List all questions in the bank
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [admin]
      summary: Add a question to the bank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /admin/questions/{id}:
    delete:
      operationId: deleteQuestion
      tags: [admin]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      summary: Get admin settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateSettings
      tags: [admin]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Student:
      type: object
      properties:
        id:
          type: integer
        nationalCode:
          type: string
        fullName:
          type: string
        grade:
          type: ["string", "null"]
        schoolName:
          type: ["string", "null"]
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
      required: [id, nationalCode, fullName, createdAt]
    LoginInput:
      type: object
      properties:
        nationalCode:
          type: string
          minLength: 10
          maxLength: 10
        fullName:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        schoolName:
          type: ["string", "null"]
      required: [nationalCode]
    AuthResult:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        isNew:
          type: boolean
      required: [student, isNew]
...
    AnalysisResult:
      type: object
      properties:
        id:
          type: integer
        recordId:
          type: integer
        ruleEngineOutput:
          type: ["string", "null"]
        interventionOutput:
          type: ["string", "null"]
        reportOutput:
          type: ["string", "null"]
        nlpOutput:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, recordId, createdAt]
    RunAnalysisInput:
      type: object
      properties:
        recordId:
          type: integer
      required: [recordId]
    Question:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        topic:
          type: ["string", "null"]
        difficulty:
          type: ["integer", "null"]
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
      required: [id, subject, text, optionA, optionB, optionC, optionD, correctAnswer]
    QuestionInput:
      type: object
      properties:
        subject:
          type: string
        topic:
          type: ["string", "null"]
        difficulty:
          type: ["integer", "null"]
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
      required: [subject, text, optionA, optionB, optionC, optionD, correctAnswer]
...
    AdminLoginInput:
      type: object
      properties:
        password:
          type: string
      required: [password]
    AdminAuthResult:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required: [success, token]
    AdminSettings:
      type: object
      properties:
        adminPassword:
          type: ["string", "null"]
        aiModel:
          type: string
        minResponseLength:
          type: integer
      required: [aiModel, minResponseLength]
    AdminSettingsUpdate:
      type: object
      properties:
        adminPassword:
          type: ["string", "null"]
        aiModel:
          type: ["string", "null"]
        minResponseLength:
          type: ["integer", "null"]
...