# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettingsUpdate {
  /** @nullable */
  adminPassword?: string | null;
  /** @nullable */
  aiModel?: string | null;
  /** @nullable */
  minResponseLength?: number | null;
}