# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface AuthResult {
  student: Student;
  isNew: boolean;
}