# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  nationalCode: string;
  fullName: string;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  schoolName?: string | null;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}